+++ src/components/animations/automate-stats.tsx
@@ -9,31 +9,30 @@ export default function AutomotateStats() {
   return (
     <section
       ref={ref}
-      className="w-full flex flex-col justify-center items-center gap-16 bg-white text-white"
+      className="w-full  relative flex flex-col overflow-hidden justify-center items-center gap-16 bg-white text-white h-[600px]"
     >
       <div
         className={cn(
-          "bg-secondary w-full flex justify-center animation-delay-2",
+          "bg-secondary z-0 absolute w-full flex justify-center animation-delay-2",
           isInView
             ? " bg-[url('/assets/pattern.png')] rounded-none bg-right bg-no-repeat bg-cover animation-delay-3"
             : "rounded-[36px] animation-delay"
         )}
-      >
-        <div className="w-full flex max-w-screen-xl flex-col gap-10 justify-center items-center rounded-3xl">
-          <h2 className=" text-4xl font-bold">
-            We automate healthcare across Africa
-          </h2>
-          <div className="w-full flex items-center justify-between">
-            {automateStats.map((item) => (
-              <div
-                key={item?.id}
-                className={cn(" flex flex-col items-center gap-3")}
-              >
-                <span className=" font-bold text-7xl">{item?.value}</span>
-                <p className=" uppercase text-sm">{item?.description}</p>
-              </div>
-            ))}
-          </div>
+      ></div>
+      <div className="w-full z-10 flex max-w-screen-md flex-col gap-16 justify-center items-center rounded-3xl">
+        <h2 className=" lg:text-[42px] font-bold">
+          We automate healthcare across Africa
+        </h2>
+        <div className="w-full flex items-center justify-between text-center">
+          {automateStats.map((item) => (
+            <div
+              key={item?.id}
+              className={cn(" flex flex-col items-center gap-3")}
+            >
+              <span className=" font-bold text-6xl">{item?.value}</span>
+              <p className=" uppercase text-xs">{item?.description}</p>
+            </div>
+          ))}
         </div>
       </div>
     </section>