+++ src/components/cards/facilities.tsx
@@ -17,14 +17,16 @@ export default function FacilitiesCard({
   bgClassName: string;
 }) {
   return (
-    <div className=" w-full rounded-lg border border-gray-100 p-7 space-y-5 hover:border-secondary hover:cursor-pointer">
+    <div className=" w-full rounded-lg border border-gray-100 p-7 space-y-8 hover:border-secondary hover:cursor-pointer">
       <div
         className={cn(" rounded-full w-fit p-3", bgClassName, textClassName)}
       >
         {icon}
       </div>
-      <h4 className=" font-bold text-lg text-gray-600">{title}</h4>
-      <p className=" text-gray-500">{description}</p>
+      <div className=" space-y-2">
+        <h4 className=" font-bold text-lg text-gray_text-700">{title}</h4>
+        <p className=" text-gray_text-500">{description}</p>
+      </div>
     </div>
   );
 }